CUDA memcpy tests.
These tests rely on a Bugle patch to support memcpy. I will update the baseline
pickle when this has filtered into the Bugle repository.

Currently memcpys are treated by Bugle as no-ops which is a source of
unsoundness. Consider the following racy kernel:

    typedef struct {
      short x;
      short y;
    } pair_t;

    __global__ void k(pair_t *pairs) {
      pair_t fresh;
      pairs[42] = fresh;
    }

which is duplicated in CUDA/fail_arrayofstruct/kernel.cu.