Generate relational pow2 candidate invariants
We look for pairs of power of two variables where one is increasing and the
other is decreasing. In this case we guess:

for n in 1,2,4,…,2^15:
  Guess candidate invariant: (up * down = n)
  for m in 1,2,4,…,2^15:
    Guess candidate invariant: (down = 0 & up = m) || (up * down = n)

Almost all of these candidates will be incorrect; so this is a very imprecise
guess.

See testsuite/OpenCL/test_relational_inference/kernel.cl for an example.