Reimplement guard variable is nonneg candidate invariant.
Old candidate was not triggering correctly.
We now move the analysis and insertion into pre-instrumentation stage.
For each loop, we find all program variables that appear in the expressions of partition variable definitions (e.g., the variables occurring in the partition variable v1 if we have a command "assume v1").
We then guess, for each variable x that 0 <= x.

This caused a previously failing test (testsuite/CUDA/misc/fail/misfail2) to verify.
By turning off inference we keep the expected failure.