Fix long-running non-negative guard generator.
The main problem was the large expressions formed by the VariableDefinition
analysis, which unrolls all variable definitions in an expression. For certain
benchmarks, with many floating point variables, an expression can become very
large. So large that when we iterate over all variables in all guard expressions
we stall the vcgen process.

Now we cut short the unrolling inside the VariableDefinition analysis. We no
longer unroll expressions that use floating point functions. See
the definition of [doNotExpandFunctions] in VariableDefinitionAnalysis.cs.