More precise guard variable detection.
We need to detect guard variables when generating non-uniform and non-negative
guard variable invariants. Given a region, we previously did this by examining
any assume expressions tagged {:partition} in the region. Then, any variable
inside an expression of this kind is a "partition variable". Finally, we
assume a guard variable is any variable appearing in the definition of a
partition variable.

We refine this by now only considering the direct successors of the header of a
region. This stops us guessing a number of unnecessary invariants. e.g.,

  for (i=0; i<100; ++i):
    for (j=0; j<100; ++j):
      ...

Previously we would generate invariants for i and j at the outer loop, when only
i is necessary.

More practically, this makes the verification of the Rightware/mandelbulb kernel
tractable again.