When giving the --params flag, one can now provide an asterisk instead of a value. This means "don't constrain this parameter".
Also, the numbers in the --local_size, --num_groups, etc. flags are now
allowed to be zero. (Previously they had to be positive.) A number being
zero in this context makes no sense, hence zero is interpreted as a
special value, meaning "don't constrain this dimension". (I wanted to
use asterisk instead of zero, but this causes typing problems.)