Properly the the no user source code location issue
If a include-blang source location is involved, it'll always be the
first in the source location list. Hence, just always adding the first
does not work. Instead, save the last include-blang source location
we see and add it to the list of records if we didn't encounter any
non-include-blang source locations.