Add an error limit option to the GPUVerify script
Fix an edge case in the handling of positive arguments while we
are here.

Unfortunately there seems be a bug in the handling of the errorLimit
argument of the Boogie version we are using: setting the error limit
to 1, I always get back two errors when using

  testsuite/OpenCL/pointeranalysistests/manyprocedures/kernel.cl

as a test case.

Closes #5