Add two more builds to the Travis matrix
These builds download an LLVM binary archive from the LLVM website
and build Bugle against these binaries. This will form the basis for
the binaries that we want to deploy.

The two builds are only build when pushing to the master branch of
the repository, as they are somewhat duplicative. Note that they
do not use GCC during the build, but use Clang instead (the compiler
the LLVM binaries were built with).